refactor: movido tela de Main
